Moving generic pytest fixtures to conftest
