+++ tests/smokescreen/test_compartment_5.py
@@ -49,17 +49,11 @@ from pages.screening_subject_search.contact_with_patient_page import (
 )
 
 from utils.user_tools import UserTools
-from utils.load_properties_file import PropertiesFile
 from utils.screening_subject_page_searcher import verify_subject_event_status_by_nhs_no
 from utils.calendar_picker import CalendarPicker
 from utils.oracle.oracle_specific_functions import get_subjects_with_booked_appointments
 from datetime import datetime, timedelta
 import logging
-
-
-@pytest.fixture
-def smokescreen_properties() -> dict:
-    return PropertiesFile().get_smokescreen_properties()
 
 
 @pytest.mark.vpn_required
